+++ src/main/scala/io/github/dexclaimation/ahql/AhqlServer.scala
@@ -9,11 +9,12 @@ package io.github.dexclaimation.ahql
 
 import akka.http.scaladsl.marshallers.sprayjson.SprayJsonSupport
 import akka.http.scaladsl.model.StatusCode
-import akka.http.scaladsl.model.StatusCodes.BadRequest
+import akka.http.scaladsl.model.StatusCodes.{BadRequest, ImATeapot}
 import akka.http.scaladsl.server.Directives.{_string2NR, as, complete, concat, entity, get, parameters, post}
 import akka.http.scaladsl.server.Route
 import io.github.dexclaimation.ahql.graphql.GqlResponse
 import io.github.dexclaimation.ahql.implicits._
+import io.github.dexclaimation.ahql.utils.HttpMethodStrategy
 import sangria.ast
 import sangria.execution._
 import sangria.execution.deferred.DeferredResolver
@@ -41,6 +42,7 @@ import scala.reflect.ClassTag
 class AhqlServer[Ctx, Val: ClassTag](
   schema: Schema[Ctx, Val],
   root: Val,
+  httpMethodStrategy: HttpMethodStrategy = HttpMethodStrategy.onlyPost,
   queryValidator: QueryValidator = QueryValidator.default,
   deferredResolver: DeferredResolver[Ctx] = DeferredResolver.empty,
   exceptionHandler: ExceptionHandler = ExceptionHandler.empty,
@@ -49,20 +51,56 @@ class AhqlServer[Ctx, Val: ClassTag](
   maxQueryDepth: Option[Int] = None,
   queryReducers: List[QueryReducer[Ctx, _]] = Nil
 ) extends SprayJsonSupport with DefaultJsonProtocol {
+
   /**
    * Route Handler for GraphQL
    *
    * @param ctx Context for the schema.
    * @return A route that already take an entity and complete with the proper results.
    */
-  def applyMiddleware(ctx: Ctx)(implicit ex: ExecutionContext): Route = concat(
-    (post & entity(as[JsValue])) { js =>
-      complete(serve(js, ctx))
-    },
-    (get & parameters("query", "variables".optional, "operationName".optional)) { (query, variables, ops) =>
-      complete(serverOverGet(query, variables, ops, ctx))
+  def applyMiddleware(ctx: Ctx)(implicit ex: ExecutionContext): Route = httpMethodStrategy match {
+    case HttpMethodStrategy.EnableAll => concat(
+      (post & entity(as[JsValue])) { js =>
+        complete(serve(js, ctx))
+      },
+      (get & parameters("query", "variables".optional, "operationName".optional)) { (query, variables, ops) =>
+        complete(serverOverGet(query, variables, ops, ctx))
+      }
+    )
+
+    case HttpMethodStrategy.EnableOnlyPost => post {
+      entity(as[JsValue]) { js =>
+        complete(serve(js, ctx))
+      }
     }
-  )
+
+    case HttpMethodStrategy.EnableOnlyGet => get {
+      parameters("query", "variables".optional, "operationName".optional) { (query, variables, ops) =>
+        complete(serverOverGet(query, variables, ops, ctx))
+      }
+    }
+
+    case HttpMethodStrategy.EnableRESTLike => concat(
+      (post & entity(as[JsValue])) { js =>
+        complete(serve(js, ctx, onlyMutation = true))
+      },
+      (get & parameters("query", "variables".optional, "operationName".optional)) { (query, variables, ops) =>
+        complete(serverOverGet(query, variables, ops, ctx, onlyQuery = true))
+      }
+    )
+
+    case HttpMethodStrategy.EnableGetForQuery => concat(
+      (parameters("query", "variables".optional, "operationName".optional) & get) { (q, v, o) =>
+        complete(serverOverGet(q, v, o, ctx, onlyQuery = true))
+      },
+      (entity(as[JsValue]) & post) { js =>
+        complete(serve(js, ctx))
+      }
+    )
+
+    case HttpMethodStrategy.EnableNone =>
+      complete(ImATeapot, GqlResponse.error("GraphQL Server disabled all routes"))
+  }
 
   /**
    * Execution Handler for GraphQL '''POST'''
@@ -70,16 +108,31 @@ class AhqlServer[Ctx, Val: ClassTag](
    * @param ctx Context for the schema.
    * @return A Future of Status Code with a response JsValue.
    */
-  def serve(js: JsValue, ctx: Ctx)(implicit ex: ExecutionContext): Future[(StatusCode, JsValue)] = js
+  def serve(js: JsValue, ctx: Ctx, onlyMutation: Boolean = false)
+    (implicit ex: ExecutionContext): Future[(StatusCode, JsValue)] = js
     .gqlContext
-    .map { case (queryAst, vars, operation) => execute(ctx, queryAst, vars, operation) }
-    .unwrapOr { e => Future.successful(BadRequest -> GqlResponse.error(e.getMessage)) }
+    .filter { case (queryAst, _, op) =>
+      !onlyMutation || queryAst.operation(op).map(_.operationType).contains(ast.OperationType.Mutation)
+    }
+    .map { case (queryAst, vars, operation) =>
+      execute(ctx, queryAst, vars, operation)
+    }
+    .unwrapOr { e =>
+      Future.successful(BadRequest -> GqlResponse.error(e.getMessage))
+    }
 
 
   private def serverOverGet(
-    query: String, variables: Option[String], operation: Option[String], ctx: Ctx
+    query: String,
+    variables: Option[String],
+    operation: Option[String],
+    ctx: Ctx,
+    onlyQuery: Boolean = false
   )(implicit ex: ExecutionContext): Future[(StatusCode, JsValue)] = QueryParser
     .parse(query)
+    .filter { queryAst =>
+      !onlyQuery || queryAst.operation(operation).map(_.operationType).contains(ast.OperationType.Query)
+    }
     .map(execute(ctx, _, variables.map(_.toJson).getOrElse(JsObject.empty).asJsObject, operation))
     .unwrapOr(e => Future.successful(BadRequest -> GqlResponse.error(e.getMessage)))